rename ProfileHeader -> Toolbar
